+++ libs/core/src/lib/utils/functions/get-string-from-hashmap.ts
@@ -1,0 +1,7 @@
+import { Hash } from '../public_api';
+
+export function getStringFromHashMap<T>(hashMap: Hash<T>): string {
+    return Object.keys(hashMap)
+        .map((key) => hashMap[key])
+        .reduce((acc, value) => (acc += value), '');
+}